pyproject.toml
[project]
name = "ai-saas-starter"
version = "0.1.0"
description = "Full-stack AI SaaS boilerplate with FastAPI, React, auth, Stripe billing, and GDPR compliance"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "structlog>=24.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "stripe>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "bandit>=1.8.0",
    "pip-audit>=2.7.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
